BC-11843 - fix delete elements
